feat: better lib interface
