+++ src/main/kotlin/parallel/ParallelMode.kt
@@ -1,0 +1,8 @@
+package org.example.parallel
+
+import org.example.image.ImageData
+import org.example.kernels.Kernel
+
+interface ParallelMode {
+    fun run(img: ImageData, kernel: Kernel)
+}
